requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.32.0
idna==3.3
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.1
nltk==3.7
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.15
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9